feat: Calculate distributions from dataset (#26)
Calculates the distributions of scores when you run the backend, it'll
generate some files that will be packaged for frontend. Meaning frontend
code won't even compile now if you haven't run the backend.

I had to add a bunch of new types so the common.rs file got a bit too
big so i also split that up.